backend/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==2.2.20
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
jaraco.classes==3.2.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==9.0.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pkginfo==1.8.3
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==37.3
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
zipp==3.8.1
